Refactored postal code into internel component
