+++ packages/lib/src/components/Boleto/components/BoletoInput/BoletoInput.tsx
@@ -1,7 +1,5 @@
 import { h } from 'preact';
 import { useState, useEffect, useRef } from 'preact/hooks';
-import { renderFormField } from '../../../internal/FormFields';
-import Field from '../../../internal/FormFields/Field';
 import Address from '../../../internal/Address';
 import { boletoValidationRules } from './validate';
 import { boletoFormatters } from './utils';
@@ -9,7 +7,7 @@ import SendCopyToEmail from '../../../internal/SendCopyToEmail/SendCopyToEmail';
 import useCoreContext from '../../../../core/Context/useCoreContext';
 import { BoletoInputDataState } from '../../types';
 import useForm from '../../../../utils/useForm';
-import SocialSecurityNumberBrazil from '../../../internal/SocialSecurityNumberBrazil/SocialSecurityNumberBrazil';
+import { BrazilPersonalDetail } from '../../../internal/SocialSecurityNumberBrazil/BrazilPersonalDetail';
 
 function BoletoInput(props) {
     const { i18n } = useCoreContext();
@@ -63,41 +61,7 @@ function BoletoInput(props) {
     return (
         <div className="adyen-checkout__boleto-input__field">
             {props.personalDetailsRequired && (
-                <div className={'adyen-checkout__fieldset adyen-checkout__fieldset--address adyen-checkout__fieldset--personalDetails'}>
-                    <div className="adyen-checkout__fieldset__title">{i18n.get('personalDetails')}</div>
-
-                    <div className="adyen-checkout__fieldset__fields">
-                        <Field label={i18n.get('firstName')} classNameModifiers={['firstName', 'col-50']} errorMessage={!!errors.firstName}>
-                            {renderFormField('text', {
-                                name: 'firstName',
-                                autocorrect: 'off',
-                                spellcheck: false,
-                                value: data.firstName,
-                                onInput: handleChangeFor('firstName', 'input'),
-                                onChange: handleChangeFor('firstName')
-                            })}
-                        </Field>
-
-                        <Field label={i18n.get('lastName')} classNameModifiers={['lastName', 'col-50']} errorMessage={!!errors.lastName}>
-                            {renderFormField('text', {
-                                name: 'lastName',
-                                autocorrect: 'off',
-                                spellcheck: false,
-                                value: data.lastName,
-                                onInput: handleChangeFor('lastName', 'input'),
-                                onChange: handleChangeFor('lastName')
-                            })}
-                        </Field>
-
-                        <SocialSecurityNumberBrazil
-                            data={data.socialSecurityNumber}
-                            error={errors.socialSecurityNumber}
-                            valid={valid.socialSecurityNumber}
-                            onInput={handleChangeFor('socialSecurityNumber', 'input')}
-                            onChange={handleChangeFor('socialSecurityNumber')}
-                        />
-                    </div>
-                </div>
+                <BrazilPersonalDetail i18n={i18n} data={data} handleChangeFor={handleChangeFor} errors={errors} valid={valid} />
             )}
 
             {props.billingAddressRequired && (
@@ -121,7 +85,12 @@ function BoletoInput(props) {
                 />
             )}
 
-            {props.showPayButton && props.payButton({ status, label: i18n.get('boletobancario.btnLabel'), classNameModifiers: buttonModifiers })}
+            {props.showPayButton &&
+                props.payButton({
+                    status,
+                    label: i18n.get('boletobancario.btnLabel'),
+                    classNameModifiers: buttonModifiers
+                })}
         </div>
     );
 }
