+++ packages/lib/src/components/Pix/PixInput/PixInput.tsx
@@ -1,17 +1,14 @@
 import { h } from 'preact';
-import { useEffect, useRef, useState } from 'preact/hooks';
-import { renderFormField } from '../../internal/FormFields';
-import Field from '../../internal/FormFields/Field';
+import { useEffect, useState } from 'preact/hooks';
 import { pixValidationRules } from './validate';
 import { pixFormatters } from './utils';
 import useCoreContext from '../../../core/Context/useCoreContext';
 import { PixInputDataState } from '../types';
 import useForm from '../../../utils/useForm';
-import SocialSecurityNumberBrazil from '../../internal/SocialSecurityNumberBrazil/SocialSecurityNumberBrazil';
+import { BrazilPersonalDetail } from '../../internal/SocialSecurityNumberBrazil/BrazilPersonalDetail';
 
 function PixInput(props) {
     const { i18n } = useCoreContext();
-    const addressRef = useRef(null);
     const formSchema = ['firstName', 'lastName', 'socialSecurityNumber'];
     const { handleChangeFor, triggerValidation, setSchema, data, valid, errors, isValid } = useForm<PixInputDataState>({
         schema: formSchema,
@@ -22,7 +19,7 @@ function PixInput(props) {
 
     // Handle form schema updates
     useEffect(() => {
-        const newSchema = [...(props.personalDetailsRequired ? formSchema : [])];
+        const newSchema = props.personalDetailsRequired ? [...formSchema] : [];
         setSchema(newSchema);
     }, [props.personalDetailsRequired]);
 
@@ -31,9 +28,6 @@ function PixInput(props) {
 
     this.showValidation = () => {
         triggerValidation();
-        if (props.billingAddressRequired) {
-            addressRef.current.showValidation();
-        }
     };
 
     useEffect(() => {
@@ -45,41 +39,7 @@ function PixInput(props) {
     return (
         <div className="adyen-checkout__pix-input__field">
             {props.personalDetailsRequired && (
-                <div className={'adyen-checkout__fieldset adyen-checkout__fieldset--address adyen-checkout__fieldset--personalDetails'}>
-                    <div className="adyen-checkout__fieldset__title">{i18n.get('personalDetails')}</div>
-
-                    <div className="adyen-checkout__fieldset__fields">
-                        <Field label={i18n.get('firstName')} classNameModifiers={['firstName', 'col-50']} errorMessage={!!errors.firstName}>
-                            {renderFormField('text', {
-                                name: 'firstName',
-                                autocorrect: 'off',
-                                spellcheck: false,
-                                value: data.firstName,
-                                onInput: handleChangeFor('firstName', 'input'),
-                                onChange: handleChangeFor('firstName')
-                            })}
-                        </Field>
-
-                        <Field label={i18n.get('lastName')} classNameModifiers={['lastName', 'col-50']} errorMessage={!!errors.lastName}>
-                            {renderFormField('text', {
-                                name: 'lastName',
-                                autocorrect: 'off',
-                                spellcheck: false,
-                                value: data.lastName,
-                                onInput: handleChangeFor('lastName', 'input'),
-                                onChange: handleChangeFor('lastName')
-                            })}
-                        </Field>
-
-                        <SocialSecurityNumberBrazil
-                            data={data.socialSecurityNumber}
-                            error={errors.socialSecurityNumber}
-                            valid={valid.socialSecurityNumber}
-                            onInput={handleChangeFor('socialSecurityNumber', 'input')}
-                            onChange={handleChangeFor('socialSecurityNumber')}
-                        />
-                    </div>
-                </div>
+                <BrazilPersonalDetail i18n={i18n} data={data} handleChangeFor={handleChangeFor} errors={errors} valid={valid} />
             )}
 
             {props.showPayButton &&
